Move utility functions to apply_modifiers.py
